tests/wpt/tests/css/css-gaps/parsing/column-rule-visibility-items-computed.html


<!DOCTYPE html>
<html>
<head>
  <meta charset="utf-8">
  <title>CSS Gaps: column-rule-visibility-items getComputedStyle()</title>
  <link rel="author" title="Sam Davis Omekara Jr." href="mailto:samomekarajr@microsoft.com">
  <link rel="help" href="https://drafts.csswg.org/css-gaps-1">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/css/support/computed-testcommon.js"></script>
  <script src="/css/support/inheritance-testcommon.js"></script>
</head>
<body>
  <div id="target"></div>
  </div>
  <script>
    test_computed_value("column-rule-visibility-items", "all");
    test_computed_value("column-rule-visibility-items", "none");
    test_computed_value("column-rule-visibility-items", "around");
    test_computed_value("column-rule-visibility-items", "between");
  </script>
</body>
</html>
